+++ config/atoms.js
@@ -1,0 +1,6 @@
+import { atom } from "recoil";
+
+const CurrentPokemon = {
+  key: 'pokemon', // unique ID (with respect to other atoms/selectors)
+  default: null,
+}
