+++ pages/index.js
@@ -1,9 +1,12 @@
 import Head from 'next/head'
 // layout
-import Header from '../components/layout/header';
-import Content from '../components/layout/content';
-import Pokemon from '../components/layout/pokemon';
+import Header from '../components/header';
+import Layout from '../components/layaout';
+import Pokemon from '../components/pokemon';
 
+import List from '../views/list';
+import PokeTable from '../views/table';
+import Gallery from '../views/gallery';
 
 //styles
 import styled from 'styled-components';
@@ -14,9 +17,38 @@ const StyledContent = styled.div`
 
 export default function Home() {
   
+  
+ 
+  // get Pokemons from APi
+  const getPokemons = async () => {
+    dispatch( getListPokemonsAction() );
+  }
+  // get Pokemon from list
+  const getPokemon = async ( id ) => {
+   
+  }
   return (
-      <>
-        <Header/>
-      </>
+    
+        <Layout>
+          {
+        screen === 'list'? <List  
+          pokemonList = {pokemonlist} 
+          getPokemon = { getPokemon }
+        /> : null
+      }
+      {
+        screen === 'table'? <PokeTable 
+          pokemonList = {pokemonlist}
+          getPokemon = { getPokemon }
+        /> : null
+      }
+      {
+        screen === 'gallery'? <Gallery  
+          pokemonList = {pokemonlist}
+          getPokemon = { getPokemon }
+        /> : null
+      }
+        </Layout>
+     
   )
 }
