+++ views/gallery.jsx
